add network devices stats to overview endpoint
